# backend/openapi.yaml
openapi: 3.0.3
info:
  title: ExamBoost Togo API
  description: |
    Backend API for ExamBoost Togo — Adaptive exam preparation platform for
    Togolese national exams (BEPC, BAC1, BAC2).

    The API exposes:
      - **Authentication** — JWT-based register / login / me endpoints.
      - **Question bank** — Paginated listing with filters (subject, exam,
        serie, competence, chapter), random sampling for exam simulations,
        admin-only creation endpoint.
      - **Learning sessions** — Records each student answer and updates both
        the SM-2 spaced-repetition state and the Bayesian Knowledge Tracing
        (BKT) per-competence mastery estimate.
      - **ML predictions** — Predicted exam score (XGBoost if >= 100
        responses available, heuristic BKT fallback otherwise) and dropout
        risk score.

    Adaptive learning algorithms implemented server-side (mirror of the
    Flutter app's local-first engine):
      - **IRT 3PL** (a/b/c parameters, Fisher information for CAT)
      - **BKT** (4-parameter model: P(L), P(T), P(S), P(G), mastery 0.85)
      - **SM-2** (Easiness Factor >= 1.3, intervals 1/6/N*EF)

    Default BKT parameters (configurable via env vars):
      - P(L) initial  = 0.10
      - P(T) transit  = 0.20
      - P(S) slip     = 0.10
      - P(G) guess    = 0.20
      - mastery       = 0.85
  version: 1.0.0
  contact:
    name: ExamBoost Togo Team
    email: contact@examboost.togo
    url: https://github.com/djabelo712/ExamBoost-Togo
...
servers:
...
security:
  - BearerAuth: []
...
paths:
...
  /questions:
    get:
      tags: [Questions]
      summary: List questions with filters
      description: |
        Returns a paginated list of questions filtered by subject, exam
        type, serie, competence, or chapter. Used by the Flutter app to
        populate the revision and simulation screens.
      security: []
      operationId: listQuestions
      parameters:
        - name: matiere
          in: query
          description: Filter by subject (e.g. "Mathematiques", "Francais")
          required: false
          schema: { type: string }
        - name: examen
          in: query
          description: Filter by exam type
          required: false
          schema:
            type: string
            enum: [BEPC, BAC1, BAC2, Probatoire]
        - name: serie
          in: query
          description: Filter by track (only relevant for BAC)
          required: false
          schema:
            type: string
            enum: [A, B, C, D, F]
        - name: competence_id
          in: query
          description: Filter by competence ID (e.g. "TG-MATHS-ALGEBRE-EQ2D")
          required: false
          schema: { type: string }
        - name: chapitre
          in: query
          description: Filter by chapter (free text)
          required: false
          schema: { type: string }
        - name: limit
          in: query
          description: Page size
          required: false
          schema:
            type: integer
            default: 20
            minimum: 1
            maximum: 200
        - name: offset
          in: query
          description: Page offset
          required: false
          schema:
            type: integer
            default: 0
            minimum: 0
      responses:
        '200':
          description: Paginated list of questions
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/QuestionListOut'
    post:
      tags: [Questions]
      summary: Create a new question (admin only)
      description: |
        Adds a question to the bank. Restricted to admin users (the first
        registered account, or any account matching `ADMIN_EMAIL`).
      operationId: createQuestion
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/QuestionCreate'
      responses:
        '201':
          description: Question created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/QuestionOut'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          description: Admin only
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
              example:
                detail: "Acces reserve aux administrateurs"
        '409':
          description: Question ID already exists
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
              example:
                detail: "La question TG-BEPC-MATHS-2022-Q01 existe deja"
  /questions/random/list:
    get:
      tags: [Questions]
      summary: Random question sample (for simulations)
      description: |
        Returns `n` random questions, optionally filtered by subject and
        exam type. Used by the simulation screen to build a fresh exam
        set on each launch.
      security: []
      operationId: getRandomQuestions
      parameters:
        - name: n
          in: query
          description: Number of questions to sample
          required: false
          schema:
            type: integer
            default: 10
            minimum: 1
            maximum: 100
        - name: matiere
          in: query
          required: false
          schema: { type: string }
        - name: examen
          in: query
          required: false
          schema:
            type: string
            enum: [BEPC, BAC1, BAC2, Probatoire]
      responses:
        '200':
          description: Random list of questions
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/QuestionOut'
  /questions/{question_id}:
    get:
      tags: [Questions]
      summary: Get a question by ID
      description: Returns the full question record including answer, explanation, and IRT parameters.
      security: []
      operationId: getQuestion
      parameters:
        - name: question_id
          in: path
          required: true
          description: Question identifier (e.g. "TG-BEPC-MATHS-2022-Q01")
          schema: { type: string }
          example: "TG-BEPC-MATHS-2022-Q01"
      responses:
        '200':
          description: Question found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/QuestionOut'
        '404':
          $ref: '#/components/responses/NotFound'

  # ─── Sessions ──────────────────────────────────────────────────────
...
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: |
        JWT access token obtained from `/auth/register` or `/auth/login`.
        Pass it as `Authorization: Bearer <token>` in the request header.
        Tokens are valid for 7 days (configurable via
        `ACCESS_TOKEN_EXPIRE_MINUTES`).
  schemas:
...
    IrtParams:
      type: object
      description: IRT 3PL parameters (a/b/c) for a question.
      properties:
        a:
          type: number
          nullable: true
          description: Discrimination parameter (typically 0.5-2.5)
          example: 1.4
        b:
          type: number
          nullable: true
          description: Difficulty parameter (typically -3 to +3)
          example: 0.2
        c:
          type: number
          nullable: true
          description: Guessing parameter (typically 0-0.35)
          example: 0.15
        calibre:
          type: boolean
          description: True if parameters have been calibrated via py-irt
          default: false
    QuestionCreate:
      type: object
      description: Payload for question creation (admin only).
      required: [enonce, reponse, matiere, chapitre, competence_id]
      properties:
        id:
          type: string
          nullable: true
          description: Optional ID. Auto-generated if absent.
          example: "TG-BEPC-MATHS-2022-Q01"
        enonce:
          type: string
          minLength: 3
          example: "Résous le système d'équations : 2x + 3y = 12 et x - y = 1."
        reponse:
          type: string
          example: "x = 3, y = 2"
        explication:
          type: string
          nullable: true
          example: "On utilise la substitution : x = 1 + y..."
        matiere:
          type: string
          example: "Mathematiques"
        chapitre:
          type: string
          example: "Algèbre — Systèmes d'équations"
        competence_id:
          type: string
          example: "TG-MATHS-ALGEBRE-EQ2D"
        examen:
          type: string
          enum: [BEPC, BAC1, BAC2, Probatoire]
          default: BEPC
        serie:
          type: string
          nullable: true
          enum: [A, B, C, D, F]
        annee:
          type: integer
          nullable: true
          example: 2022
        type:
          type: string
          enum: [calcul, ouvert, qcm, vraiFaux, redaction]
          default: ouvert
        choix:
          type: array
          nullable: true
          description: Choices for QCM questions (required if type=qcm)
          items: { type: string }
          example: ["3", "5", "7", "9"]
        points:
          type: integer
          nullable: true
          minimum: 1
          maximum: 20
          example: 4
        irt:
          $ref: '#/components/schemas/IrtParams'
    QuestionOut:
      type: object
      description: Full question record (serialised for API responses).
      properties:
        id:
          type: string
          example: "TG-BEPC-MATHS-2022-Q01"
        enonce:
          type: string
        reponse:
          type: string
        explication:
          type: string
          nullable: true
        matiere:
          type: string
        chapitre:
          type: string
        competence_id:
          type: string
        examen:
          type: string
          enum: [BEPC, BAC1, BAC2, Probatoire]
        serie:
          type: string
          nullable: true
          enum: [A, B, C, D, F]
        annee:
          type: integer
          nullable: true
        type:
          type: string
          enum: [calcul, ouvert, qcm, vraiFaux, redaction]
        choix:
          type: array
          nullable: true
          items: { type: string }
        points:
          type: integer
          nullable: true
        irt:
          $ref: '#/components/schemas/IrtParams'
    QuestionListOut:
      type: object
      description: Paginated list of questions.
      properties:
        items:
          type: array
          items:
            $ref: '#/components/schemas/QuestionOut'
        total:
          type: integer
          description: Total number of questions matching the filters
          example: 64
        limit:
          type: integer
          example: 20
        offset:
          type: integer
          example: 0
...
    Error:
      type: object
      description: Standard error response (FastAPI format).
      properties:
        detail:
          oneOf:
            - type: string
              example: "Utilisateur 550e8400-e29b-41d4-a716-446655440000 introuvable"
            - type: object
              description: Validation error detail (FastAPI wraps multiple errors)
            - type: array
              items: { type: object }
  responses:
    BadRequest:
      description: Bad request — validation error or malformed payload
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
          example:
            detail: "niveau_scolaire doit etre dans ['2nde', '3eme', '1ere', 'Terminale', 'tle']"
    Unauthorized:
      description: Unauthorized — missing or invalid JWT token
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
          example:
            detail: "Token invalide ou absent"
    NotFound:
      description: Not found — resource does not exist
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
          example:
            detail: "Question TG-BEPC-MATHS-2022-Q99 introuvable"
...